reorganize so that numerical arrays are just one more interpretation
