+++ uproot/interp/jagged.py
